ModelPipeline/AutoPromptRefinement/requirements.txt
mysql-connector-python==8.3.0
google-auth==2.29.0
google-api-python-client==2.125.0
PyYAML==6.0.1
crewai==0.26.10
